Update hetzner-debian12-zfs-setup.sh
fix for 
===========remove unused kernels in rescue system=========
Reading package lists... Done
Building dependency tree... Done
Reading state information... Done
E: Unable to locate package linux-headers-6.11.3
E: Couldn't find any package by glob 'linux-headers-6.11.3'

+++ hetzner-debian12-zfs-setup.sh
@@ -491,8 +491,12 @@ clear
 
 echo "===========remove unused kernels in rescue system========="
 for kver in $(find /lib/modules/* -maxdepth 0 -type d | grep -v "$(uname -r)" | cut -s -d "/" -f 4); do
-  apt purge --yes "linux-headers-$kver"
-  apt purge --yes "linux-image-$kver"
+  if dpkg -l "linux-headers-$kver" 2>/dev/null | grep -q "^ii"; then
+    apt purge --yes "linux-headers-$kver"
+  fi
+  if dpkg -l "linux-image-$kver" 2>/dev/null | grep -q "^ii"; then
+    apt purge --yes "linux-image-$kver"
+  fi
 done
 
 echo "======= installing zfs on rescue system =========="
